requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
nltk==3.8.1
python-dotenv==1.0.1
regex==2023.12.25
six==1.16.0
tqdm==4.66.2
Werkzeug==3.0.1
